Add webhook URL verification for REST API
When a webhook URL is provided via the REST API, the `verifyWebhookURL` function checks that it is valid URL before saving it to the database (if valid) or throwing an error (if invalid).

To verify the URL is valid, this JSON payload is sent to the client

```
{
    "applicationID": "xxxxx-xxxx-xxxx-xxxxx",
    "event": {"message": "Please respond with a HTTP 200 status code."},
    "timestamp": "1689259011623",
    "type": "url-verification"
}

```
A timestamp and signature is sent in the header, just like we do when sending webhook payloads, so clients can verify the request's authenticity.

We then wait for a response with status code 200 from the client app. The response should be received within 3 seconds, otherwise the verification fails (I chose 3 seconds becuase that's the time limit Slack set for their URL verification.).

In the create application endpoint, the new application is first created but if URL verification fails, then the create is rolled back and not written into the databasse.


Reviewers: yeo-yeo, dmmiller, flooey, Netceer

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/5506

monorepo-commit: e9be2fb0cb8112eece5e5d85df6caeccd7d08881