Refactor options handling in init (#5548)
Do some refactors of option handling in `CordSDK.init`.  This is in
preparation for introducing a new function that only changes the
options, rather than requiring re-init, since that's both more clear
and means you don't have to keep your access token around if all
you're going to do is flip options.

Break checking and cleaning options out into its own method.  As part,
have this be the one place we canonicalize options for backwards
compat, and everything else uses the canonicalized version.

Break out a separate function that just sets the options and
rerenders.  This will be used by the `setOptions` function in the
future.

Test Plan:

Testbed still inits just fine, rerunning init again with different
options but the same access token causes everything to rerender but
not reconnect the websocket.

monorepo-commit: 847182420040b045620e6f67c27e2cb02962c72d