Add reading message attachments (#5602)
Add info about message attachments to the message data structure in
REST APIs, JS APIs, and message-added webhook.

Test Plan:

Called REST API to fetch a single message with an attachment, it
worked.  Also fetched a message with two attachments.

Called REST API to fetch the thread that contained both above
messages, they came through.

Called thread.observeThreadData() on thread with above messages, it
returned the attachments.  Added a new message on that thread with an
attachment, it came through to the callback.

Logged webhook events and sent a message with two attachments, they
showed up in the payload properly.

monorepo-commit: 0568546a684d44ca95c1a8e2b535d2f5853000fb