Enabling mark as seen messages for ThreadedComments
I am adding this mitigating fix where we only mark the first message as seen if it is a
single message. This is to avoid a case where we can't mark a thread as seen if it doesn't
have replies. If it does, we can simply wait for the user to view the replies to mark it as
seen.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/8dbec07d-c2cf-4b69-a0a4-61471c83cdc9



Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/5633

monorepo-commit: e5e637e1da5ffe8f2f6248b8a321ffa9f743ec2c