Add first version of translation framework (#5688)
Add an initial version of a translation framework to let people
replace strings in our client code.  This is more of an RFC than a
full implementation, but the intention is that it could be committed
and deployed as-is to test it out.

The key use case is that a caller can provide a `translations`
configuration option with an appropriate form and we will use that to
display strings.  We supply our own strings (and could include them in
open source code if we wished) for everything, so if you have no
interest in changing the strings the behavior won't change.  There's
also a `language` option that changes what language it uses for
strings.

Things that aren't in this diff:

* I haven't extracted every single bit of text in the product.  That
  is a bunch of mechanical work and I want to get the approach checked
  out first.

* The handling of timestamps isn't at all complete.  We'd need to be
  much richer here, particularly since we use these as part of a
  fragment (and thus should supply the whole fragment as a translation
  unit), and half of them are dayjs format strings.  But this gives us
  a start.

* This doesn't work for any messages that are constructed on the
  server, such as notification text or thread-resolved messages.  I
  think the best approach will be to not actually manufacture those
  and instead send a translation key to the client and have it
  assemble the message.  But I don't think that's needed for V1.

* We should document what user properties are available for
  interpolation and make sure we supply those.  Right now it's just
  using UserFragment, which isn't quite right.

Test Plan:

Used `CordSDK.updateOptions({translations: {en: ...}})` to set some
translations, they were reflected in the UI.

monorepo-commit: 0bbfc30749d44c78c9326109c010de88706f3d73