RFC: Move API types to @cord-sdk/types (#5735)
This is an initial attempt to merge the types from
`@cord-sdk/api-types` into `@cord-sdk/types` to make things simpler.
This moves only the thread and message types, but those are the most
involved, so if it's acceptable for those it should be good for the
others.

The naming convention I used here is to prefix types with `Client` or
`Server` depending on if it's used in the JS or REST API,
respectively.  That seems pretty clear to me, but interested in
opinions.

I continued to export the modification types from
`@cord-sdk/api-types` so that all the validation machinery continues
to work, but that could get merged in as well once all the types are
moved.

Test Plan:

All tests still pass.

Docs pages for thread and message REST APIs look almost identical,
except that the types are improved on the core data (eg,
message.attachments has its subfields instead of just being typed as
`object[]`).

monorepo-commit: 9d468425092cb909a983055b59a975d046e8ecc2