Logging open-source React component instantiations
I noticed that we don't log the ThreadedComments instantiations, as it is an open-source component and it doesn't go through our normal render process. To solve this I have added a set of strings on the CordSDK object called `__CORD_OPENSOURCE_COMPONENTS`, which we can add component names to and log when they are rendered.

Currently, when we log component instantiations in our React tree, we deduplicate component names. So if a page renders a message component 3 times, we only log it once. Because we only pass strings to `__CORD_OPENSOURCE_COMPONENTS`, I made it a set so they are immediately deduplicated.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/54887172-aa51-411d-90ca-004e9f336012






Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/5691

monorepo-commit: da6fa9ecc09c3592f1ca9bcaf8dddcb921eed7ef