Exposing more data in users REST API endpoint (#5674)
## Summary

We're using this endpoint in the console. We were returning the id
already, but because we weren't checking that the return data was of the
right type, the id field was not documented. Now it's typed and
documented, and we're adding also the created timestamp, so that we can
show it in the console.

## Test plan

The PR that will be on top of this one makes the API request and uses
the data, where you can see it working.

monorepo-commit: edd54ff2e45ec922170417e0e6e538981453d492