Adding className prop to ThreadedComments
ThreadedComments doesn't accept classNames at the moment. This also makes it hard to
increase the specificity of CSS on the container div. Adding a prop so developers
can input any className they want to the container div of ThreadedComments.

I used the name `className` to follow the React naming of adding classes to a div. I think it makes
sense to follow the pattern since this prop behaves exactly the same way.

Note: I'm not sure it make sense to document this for the aforementioned behavioral reason. Let me
know if you think we should.

Test Plan:
Added the className "foo" in the testbed.
![image](https://github.com/getcord/monorepo/assets/102963329/738f7ff8-8e82-480b-b968-9abd17971853)


Reviewers: flooey, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/5833

monorepo-commit: dddad4abd1f034541d89699a5e80cecb2f1d1111