Add useMemoObject hook (#5884)
Add a hook to our opensource package that allows memoization of
nontrivial objects.  Several of our hooks take such objects as
arguments, and customers are likely to supply them as object literals,
meaning they would trigger reevaluation of any hooks that have them as
a dependency on every cycle.  This allows memoization of those, so
that despite changing identity on every render they don't trigger
extra work.

Test Plan:

Specified `(a, b) => a === b` as the isEqual argument in `useUserData`
and loaded the testbed, it infinitely rerendered.  Replaced it with
the correct value, and it did not.

monorepo-commit: 7cf2fe32f313a6ef179e869e613a69b565727fbe