Use useMemoObject in all JS API hooks (#5887)
Replaced all the memoization of options and locations with
useMemoObject, including one place we weren't memoizing the options.

I didn't replace useMemoLocation, which is used in the Annotation API,
because the typing wasn't straightforward and I didn't want to spend
time on it.

Test Plan:

Loaded all sections of the testbed, which exercises most of these
APIs.  Nothing infinitely rendered.

monorepo-commit: f8b4e4e8442715dd25413e981269ca73b5daabaf