Ability to set message to 'editingMode'
One of the things that came out of building clack - we removed the options
menu and replaced it with a custom one, with our own edit buttons. this
meant that we could use the api to update a message but still needed
a 'editing' composer to appear in place of that message.


Reviewers: jwatzman, AlbertQM, flooey

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/5772

monorepo-commit: 194b9a3ff2bcc5a6ce240881d453d815251231c9