Adding pagination to users rest api endpoint (#5824)
## Summary

We're loading all users at once. This is slow and expensive and we
should have pagination.

We could upgrade to v2, but because this API is not being used
externally, we can change the API as long as we land the console change
at the same time.

In this case, I think having offset pagination would work ok too, but
because in threads we need to add cursor based pagination, we might as
well just have all pagination being cursor based.

## Test plan
We didn't have any tests for the list endpoint, so I added one, that
makes a call and uses the returned cursor to query for more.

monorepo-commit: 96a5b476ad05bd4ee305ef3d578ae99c65743039