Removing "reactions" object from the ServerCreateMessage and ClientCreateMessage types
To add reactions to a message created from the REST API, we want developers
to use the `addReactions` object. Our `ServerCreateMessage` object also contains the
`reactions` object, from the `CoreMessageData`. I could not find any usage of the
`reactions` object in the create message code, but I could find usage of the `addReactions`
object. So I am assuming it is just a typo that it is not being omitted in this type.

No need to change the `ServerUpdateMessage` type, as it extends the `ServerCreateMessage`
type.

Same applies for the ClientCreateMessage and ClientUpdateMessage types. I will be adding the
add/remove reaction objects on the client in a subsequent PR.

Test Plan:
eyes + no TS errors + docs look okay.


Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/5891

monorepo-commit: 3d3fbdb41599cad1f73b75d4fae73bdd0a44864c