Adding reactions to the send/update message js API
Adding the `addReactions` to the sendMessage API and `addReactions` and `removeReactions` to the updateMessage API. Since we shouldn't ask for a userID and a timestamp when adding a reaction (as we do in  the REST API), I made the `ClientAddReaction` and `ClientRemoveReaction` types simply strings and made the API accept a list of reactions instead of objects. The UserID and timestamp are filled server-side, in the resolvers.

Note: The ordering of the docs is a bit weird - add and remove reactions are not one after the other etc. - but it is something out of our control at the moment. Not sure if this is a problem right now.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/ada3a6a7-871d-49fb-a045-cefa10a622c6




Reviewers: jwatzman, maylynn-ng, flooey

Reviewed By: maylynn-ng, jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/5888

monorepo-commit: 848d0b3d4616e955e7acc6b5832cffcc5dbcd088