Bugfix: Expand replies on the first reply of ThreadedComments
When sending the first reply to a ThreadedComments thread, the replies
immediately appeared as collapsed. That was because when we first reply, we
don't change the state of `showingReplies` to `true`. This did not apply in
any other reply after the first, as we had to click on the thread facepile,
which set `showingReplies` to true first.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/5fba3592-97ee-4b49-bd5e-a9b7fca07437



Reviewers: AlbertQM, flooey

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/5938

monorepo-commit: 063aa03f3db9e37ede1e51ba1530702e1f352996