List all messages rest api (#5881)
## Summary

This is a "fun" one... We need to show all the messages created in a
given application in the console. There's no endpoint that does this, so
we need to add a new one. And because there will potentially be a lot of
messages, we need pagination.

In users, I managed to ignore the timestamp in microseconds problem,
which I cannot do anymore. Because we're sorting the messages by
creation timestamp, part of the cursor has to be the creation timestamp
of the last item returned. In order to get the actual value, not the
truncated js version, I've added a new virtual column to MessageEntity,
that returns in text format the creation timestamp, including the
microseconds. This is in such a format that we can do a comparison in
the query, bypassing the js limitations.

Another weird thing I had to do is do some eager fetching for the thread
a message belongs to, so that we can then fetch more information. This
is fairly easy but messes up a bit with typescript. If you think it's
not nice, I can find another way of doing it.

## Test plan

Added a new test that should cover two requests to the endpoint, making
sure that pagination works as it should.

monorepo-commit: 4b49f0464f8d56094e482158fa2039f8016cb1f2