Moving styleProps and all atomicClasses to the opensource package
We want to move the Icon component to the open-source package so we can use icons in our
open-source components. The Icon component uses styleProps and styleProps uses all of the
atomic classes, so the first step is moving these together.

The one file not used by styleProps but is in the atomicClasses folder is icons.css.ts. I didn't think
it made sense to keep the folder internally for a single file, so I moved this as well.

Test Plan:
Clicked around in the testbed, made sure nothing is broken.


Reviewers: lazybean, flooey

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/5953

monorepo-commit: c3df3707bf95f6a96584ac2891801970ba355d40