Creating an EmptyStateWithIcon.classnames file in the open-source package
We want to move the Icon component to the open-source package so we can use icons in our
open-source components. Icon.css.ts needs some classnames from the `EmptyStateWithIcon`
component. This PR creates an `EmptyStateWithIcon.classnames` file and puts it in the
open-source package so we can use it wherever we want.

I have put the file together with our React components, but in a sub-folder called `helpers`
to keep it separate from our public components, just in case a developer is looking at our
package and comes across this file - which may be confusing.

Test Plan:
`EmptyStateWithIcon` is used in the NotificationsList. I added this component
in the testbed - deleted all notifications. The empty state appeared with the correct styles.


Reviewers: flooey, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/5956

monorepo-commit: b05bc685afac0580892636fc2d73ad2f2ebe0f8b