Expose annotation logic as string encode/decode functions
This might be useful to folks, and I want to use it in
`MultipleCursors`, so might as well package it up. The underlying code
is pretty complicated and messy, so actually opening up the underlying
code doesn't seem super useful, but we can provide an API around it.

There's a bunch of iframe handling which I haven't wired up to this API,
but we can add that later if we want.

I want to use this to make `MultipleCursors` better, but it also could
be useful to position a `Pin` (instead of using `FloatingThreads`) or
potentially other things we haven't thought of.


Reviewers: AlbertQM, flooey

Reviewed By: AlbertQM, flooey

Pull Request: https://github.com/getcord/monorepo/pull/6114

monorepo-commit: 676d6be331ad9337ba2e89f18689b238e72e3c8a