Fix MultipleCursors destructuring check (#6125)
The destructuring check in MultipleCursors was slightly wrong, in that
it used `Record<never, never>` as the type.  It turns out everything
is a `Record<never, never>`, because the first `never` says you're not
allowed to look up any properties, so it's vacuously true.  Switch to
`Record<string, never>`, which says that no matter what string you
look up, it doesn't work, which only matches the empty object type.

Test Plan:

Tried adding a new property to MultipleCursorsReactComponentProps.
Before this diff, it succeeded, after it fails with a compile error.

monorepo-commit: a1687e119e818c8daab1cd56fd7979e6f16a3d1c