Fix TS types for message editing (#6127)
The TS types for various message editing APIs had gotten a couple
readonly fields into them because we forgot to exclude them, even
though they didn't do anything.  Remove those.

Also add verification in the destructuring that we're handling all the
fields of these types, so that if we add a readonly field and don't
exclude it from the modification operations we'll be notified.

Test Plan:

Viewed docs, the readonly fields no longer appear.

monorepo-commit: 6b2afec47ac926a04e8b20e9cab78f9db9d6f7cf