Remove/replace lodash
lodash was designed for a world a decade ago, where JS wasn't nearly as
advanced as it is today, and reflects that world. It's still got a ton of
useful utility functions, but their implementation leaves a lot to be desired.
It's a gigantic library code-size-wise, because the implementations try to deal
with now-ancient browsers with limited JS. Worse, the way it was written isn't
amenable to tree-shaking, and it has a frankly bizzare split between its cjs
and esm versions that cause us huge build system headaches. It also looks like
lodash is no longer under active development, with no commits in the last
couple of years.

But worst is that it's so easy to accidentally include the whole darn thing and
inflate the size of `sdk.latest.js` -- which #5845 did, by ~23k *compressed*
bytes (~4%!), and I only noticed because I just so happened to be looking at
some logging I added months ago.

And there isn't even a good way to fix it due to our build system
issues. The default lodash package is cjs-only, which is fine for our
internal build since we compile from esm to cjs anyway. But for our OSS
packages, we provide both esm and cjs versions, so that doesn't work.
(It *might* without the esModuleInterop issues we have, but that's an
entire other can of worms.) There is an esm version of lodash, but it's
a completely separate package, and using that breaks our internal build.
I could dig into what's going on here and try to solve it -- it probably
is solvable -- or we could move away from the only package in the entire
world which has caused us this sort of grief.

This PR replaces all of our usage of lodash with other things.


Reviewers: #infra, flooey, elenaperezrioja

Reviewed By: elenaperezrioja

Pull Request: https://github.com/getcord/monorepo/pull/5983

monorepo-commit: 86b7f6e096204ed757a3a5706570f218d1feae8d