Add beforeCreateMessage hook (#6116)
Add a hook that runs before a message is created in the composer to
allow the developer to hook into it and change things.

The interface here is reasonable simple, just provide a
ClientCreateMessage (the same object they would use to create a
message) and accept one back.  If they return null or undefined, skip
sending the message.  We also provide a context object, which
currently only has the thread ID, but could have other things
(location or whatever) if it proved useful.  They can also return a
promise, in case they want to make an API call or something.


Test Plan:

```
CordSDK.updateOptions({
    beforeMessageCreate: (create, context) => {
        create.content.push({type: "p", children: [{text: "Sent from my iPhone"}]});
        create.addReactions = ['💡', '😲', '🤩'];
        setTimeout(() => {
            CordSDK.updateOptions({
                beforeMessageCreate: (create, context) => {
                    create.iconURL = 'https://upload.wikimedia.org/wikipedia/commons/thumb/6/6b/Noun_Project_lightbulb_icon_1263005_cc.svg/205px-Noun_Project_lightbulb_icon_1263005_cc.svg.png';
                    create.type = 'action_message';
                    return create;
                }
            })
        }, 100);
        return create;
    },
});
```

Send two messages.  The first had two reactions on it and additional
content, the second was an action message with a lightbulb icon.

monorepo-commit: 1348a881ba844fb7670496081a9cfd0710d465a1