Use firstMessage instead of newThread in beforeMessageCreate (#6193)
The `newThread` property was checking if we were going to create a
thread, but in fact for a bunch of components, we end up creating the
thread as soon as the component is rendered.  In most cases, the
relevant piece of information is whether this is the first message
rather than if it's a new thread (which are sometimes different
because of 0-message threads).  So tell them about whether it's the
first message instead.

Test Plan:

Added a callback that logged the context and sent some messages.
firstMessage was correct when sending a message via `Composer` (which
creates a thread) and via `Thread` (which already has a 0-message
thread).

monorepo-commit: 7e102b322ac21093918580c2605c53892f7cf5e0