Privacy: fix organizationId option to useLocationSummary
I noticed that `super-secret-channel` in Clack was always showing as
unread. The JS API gives the correct data, but the React API doesn't
thread it through properly -- it was sending `organizationId` directly
instead of inside a `filter`. Instead of adding that, move to
`useMemoObject` which makes this a lot simpler and less error-prone.

Kind of annoying that TS doesn't help you with this since it isn't a
literal!

Test Plan:
Difficult to test with Clack, but it looks right now.


Reviewers: flooey, yeo-yeo

Reviewed By: yeo-yeo

Pull Request: https://github.com/getcord/monorepo/pull/6215

monorepo-commit: c8fc9a43fdcf18a2b2318aabb10b0a8be46beeae