Add file upload JS API (#6255)
Add a JS API to upload files.  This will be used for attaching files
to messages.

We use the same two-step approach we use in the composer: first we
create the file object on the backend, then we upload the file to
storage.  All of this is better managed by the SDK rather than forcing
the caller to do both parts, so we have one function (`uploadFile`)
which returns a Promise which resolves once the first step is
complete.  That Promise contains a property with a second Promise that
will resolve when the upload finishes, so that callers can track that
if they so desire.  However, the ID is available to be used for
attaching to messages as soon as the first step is complete, so I
expect most callers will just wait for the first Promise to resolve
(which should be fast, and doesn't depend on the file size) and then

I decided to accept a Blob for the file data, since that will work for
file inputs, drag-and-drop, and lots of other situations.  If we
wanted, we could support other inputs in the future (eg, data URLs),
so I named the input field `blob` rather than something more generic
so there's room for other types.

I'm not documenting this yet since until message attachments exist,
it's useless.

Test Plan:

Added a file input to a page, used it to select a file, then ran

```
await CordSDK.file.uploadFile({name: 'test.png', blob: document.getElementById("testFileInput").files[0]});
```

In the network tab, I could see that the file was properly uploaded,
and it showed up correctly in the DB.

monorepo-commit: 5711e1fda150f068e7741d3852439dd5417da847