Add a REST endpoint to call add a webhook programatically
This adds a new REST endpoint that can be called to add a webhook url and set of events that you'd like to subscribe to.  It also adds an endpoint to unregister that url.


Test Plan:
./scripts/testbedcurl.sh -X POST "https://localhost:8161/v1/applications/b6501bf5-46f7-4db7-9996-c42dd9f758b0/webhooks" -d '{"url":"https://foo.com/", "events":["thread-message-added"]}' -H "Content-Type: application/json"

./scripts/testbedcurl.sh -X DELETE "https://localhost:8161/v1/applications/b6501bf5-46f7-4db7-9996-c42dd9f758b0/webhooks/<id_returned_from_above>"

And checked database in between to see it was created and then also removed

Reviewers: yeo-yeo, flooey, jwatzman

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6249

monorepo-commit: 41f525e6b511af5ae4f3bdc5e3dc0c75ec203458