Replace radash isEqual with fast-deep-equal
The isEqual in radash doesn't handle ES6 Set/Map, and we were relying on
the lodash implementation handling them properly (which was replaced in
#5983). Several other deep-equal implementations don't handle that case
either -- but it turns out that we (at least internally) were already
also using `fast-deep-equal`, and the `/es6` submodule of it *does*
handle this case properly. So let's switch to that everywhere.

This runs afoul of the `esModuleInterp` build system crap again, which I
was trying to be done with in #5983, but at least only in one place and
with a much smaller and less error-prone package than lodash.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6263

monorepo-commit: dd66ca6ada4f931b58f1b65d8b9c3e8ba3cd12f8