Creating separator classnames file in the opensource package
We will be needing the separator classname in #6175 to apply specific styles on it. It is currently an internal component, so I am creating a classnames file to be able to access the variable in the opensource package.

Test Plan:
Separator appears correctly in the optionsMenu component. Also tested in #6175


Reviewers: AlbertQM, lazybean, flooey

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/6182

monorepo-commit: 4e164fb8294348d814f683cabf4d075e443e43ed