i18n: Handle timestamps better (#5901)
Add handling of timestamps into the translations system properly.  The
previous system had a lot of difficulties: it used fixed message
formats, it had a few properties that changed which strings were used
in very small ways (such as changing capitalization), and in presence
it was then used in string concatenation to produce an actual user
string.

Instead of using location + short/long and adjusting the formatting
based on that, make the minor variations in text part of the
translation strings, and use two different functions (that access two
different sets of translation strings), one for relative timestamps
and one for absolute timestamps.

As part of this, remove the `verbose` property from the
`cord-timestamp` component.  It's easier and more useful to supply
small modifications to the timestamp display through translations
rather than having a whole property whose entire change is just to
drop two or three words.

Fixes E-2880

Test Plan:

Compared the strings to equivalent strings shown for the same dates in
production for all of presence, notifications, and messages, every one
I checked looked the same.

Used `CordSDK.updateOptions` to update the translation string for
presence.  It updated the string used in the presence widget, but not
the same string used in the message header.

monorepo-commit: d36d3d2000c27f5097eefade21ea7da1561d9bf7