Add location and organization to notifications filters
This adds the filters to the `NotificationList` component and  the javascript
API. They both use the same function in the background to get the data and
mark_notification_as_read mutation to mark the notifications as read.

I noticed that the subscription doesn't return filtered data, even for the
existing metadata filter, so I'll get that working first in a separate diff.

Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6205

monorepo-commit: 1e68f8d4dd88fe3575dadf9656a8a1576775f9bf