Adding a resolved prop to ThreadedComments
Allowing developers to choose whether they want to show resolved threads. The new prop called `displayResolved`
has 4 options: `interleaved`, `separate`, `resolvedOnly` and `none`. Which are hopefully self-explanatory.

When `separate` is chosen, a tab on the top of the component appears with labels `Open` and `Resolved`,
which will filter the list of threads based on which button is active. Also, the composer will be hidden when resolved threads are shown, since there is no way to actually create a new resolved thread from the composer.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/3034b768-15ea-47f2-a4c3-e90601a30120

Interleaved:
<img width="385" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/92a8d5a0-9f6b-4d8c-b96b-547e45d20640">




Reviewers: dmmiller, flooey, AlbertQM, jwatzman, lazybean

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/5967

monorepo-commit: 7e049fa9285838c0eaf29ac6662f10b477f38e5d