Privacy: make organization nullable in ViewerUserData / IdentityContext
The first step towards making org ID optional in the access token is
making our code happy with the previously-omnipresent-and-global org not
being present. This marks our context types as nullable, and fixes the
resulting errors.

The OSS type `ViewerUserData` is also updated, because if there's no org
it's optional there too. This is in some sense a BC break, but in
practise is fine -- we won't pass null back unless you give us a null
org in the first place, which is *impossible* right now, and so this
won't change behaviour unless folks change on their end first.


Test Plan: TS is happy.

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6317

monorepo-commit: ed885c2f451463bc568b704bea923c9d50e848d5