MultipleCursors: should be position: fixed
Its coordinate are (and always have been) relative to the viewport, not
the page / any particular ancestor, and so `position: fixed` is the
right thing. (I guess most/all of our testing has been on the testbed
which doesn't have a positioned element above `MultipleCursors`?)


Test Plan:
Put `<MultipleCursors>` in a random place on the page, with a positioned
ancestor. The cursor shows up in the right place still.

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6347

monorepo-commit: 3b4462de30fd1d5d0b251a43abbcd462202298a9