Privacy: remove server-side requirement for organization_id in auth token
This removes the server-side checks enforcing that you must have an
`organization_id` in the auth token. It also makes that field optional
in the OSS types, which is a little unfortunate at this early stage, but
all our validation logic is based on that type so this seemed the best
way to do it -- unless you pass a feature flag, the field is still
required and you'll get an error back if you omit it.

There's obviously a lot more work to be done to make a token without an
org actually *work*, but this is a necessary first step.


Test Plan:
Sign a token without an org ID and try to do an SDK init. It swiftly
explodes, but does so *after* the server request init code now instead
of *before* it.

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6335

monorepo-commit: 80265b574d18604d864ec3e526d5a031159a2746