Privacy: client should pass along user ID even if org ID is missing in token
These checks should be separate, so we can pass along a user ID even if
the organization ID is missing. Furthermore, missing an org ID isn't
necessarily an error -- this is checked server-side based on a feature
flag and some other logic -- so let the server deal with it. See #6335.


Reviewers: yeo-yeo, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6340

monorepo-commit: a6d6d8bcb9b80bbdefd3bd8f9be5d193b53792d4