Filter: add 'location' to filter prop on threadList and observeLocationData api
We have a lots of filters in multiple places in a bunch of different ways.
This is the first of a few diffs to streamline how users pass in filters.

In this PR we're cleaning up(maybe slightly messing it up with the duplicate
location field) the observeLocationData api and threadlist component. This
was the simplest of the API as it already had the filter object passed
through with the metadata. all we needed to do was pass locations as well
as just handle it the same way we are already handling the existing location
prop.

We can come back later and remove the property in the main props, but for
now i think it’s fine to have duplicates


Reviewers: jwatzman, flooey

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6281

monorepo-commit: 209654786b04cae4e255ad05de2ec023f2565275