LiveCursors: rename MultipleCursors to LiveCursors
We decided this was a way better name.

This moves all of the code to the `LiveCursors` name. There is a
potential BC-breaking change that I want to make in the next week or so,
so I've exported it from our React package as `experimental.LiveCursors`
(though as per our convention I haven't prefixed the `cord-live-cursors`
webcomponent since folks don't really have a way of finding those names,
and the one customer we know is using this is using React anyway). I'll
move it out of experimental once we've stabalised the semantics and are
ready to release the docs, which shouldn't be too long. At that point
I'll also mark `MultipleCursors` as deprecated.

For BC, I've kept exports under the `MultipleCursors` name. Right now
they are just simple re-export aliases, but when/if we break BC on
`LiveCursors`, they can become shims to tell `LiveCursors` to use the
old behaviour. (Similarly for `cord-multiple-cursors`.)


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6446

monorepo-commit: 2c94fee7154342818fe13b8829bbaf7f038189e8