LiveCursors: rename MultipleCursors to LiveCursors (take 2)
Originally landed as #6446, reverted as #6456 due to breaking SDK init.

The issue was that I defined two custom elements to both use the same
constructor, which is a violation of the HTML custom elements spec
section 4.13.4, "element definition" rule 4:
https://html.spec.whatwg.org/multipage/custom-elements.html#custom-elements-api

Apparently this is so that we can have a `getName` function which maps a
constructor to its element name, and so the mapping has to be bijective.
I think that this is a mistake in the spec, if anyone cares what I think
(they don't).

Work around it by defining a new class for legacy
`cord-multiple-cursors` which just extends the superclass. Would likely
need to do this anyway for BC purposes.

Original PR message:

---

We decided this was a way better name.

This moves all of the code to the `LiveCursors` name. There is a
potential BC-breaking change that I want to make in the next week or so,
so I've exported it from our React package as `experimental.LiveCursors`
(though as per our convention I haven't prefixed the `cord-live-cursors`
webcomponent since folks don't really have a way of finding those names,
and the one customer we know is using this is using React anyway). I'll
move it out of experimental once we've stabalised the semantics and are
ready to release the docs, which shouldn't be too long. At that point
I'll also mark `MultipleCursors` as deprecated.

For BC, I've kept exports under the `MultipleCursors` name. Right now
they are just simple re-export aliases, but when/if we break BC on
`LiveCursors`, they can become shims to tell `LiveCursors` to use the
old behaviour. (Similarly for `cord-multiple-cursors`.)


Reviewers: flooey, yeo-yeo

Reviewed By: yeo-yeo

Pull Request: https://github.com/getcord/monorepo/pull/6458

monorepo-commit: ebda242f89080bc9cadfa05216a640923c43d129