Disabling the resolving of threads when we do not display resolved threads in ThreadedComments
Hiding the `Resolve` button from the ThreadedComments threads if we are not showing the resolved tab. Also hiding the reopen buttons when only showing resolved threads (these include the header reopen button, reopen-to-edit options menu  item as well as the separator which was no longer necessary and reply composer in resolved threads).

Since there is no way to access resolved/open threads in each configuration, there is no point in allowing users to resolve/reopen them. This also becomes useful in the case where ThreadedComments is used as a comment section. There is no notion of resolving a Youtube/blog post comment - for example. Because this is css though - it is very easily overridable :)

The implementation simply adds a css class to add `display: none` in the relevant menu item css. This is not the most robust solution, as users can still resolve/reopen a thread via the client side API. But the alternative solution required quite a lot of refactoring and a ton of work in making our options menu buttons more flexible, so we went with this one, for now.

Note: I have mapped the classNames to the UI names of the tabs, rather than the prop. I think it makes more sense from a developer perspective, but let me know if you disagree.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/76ee8569-4593-48fe-85d6-a8f387f786df




Reviewers: AlbertQM, lazybean, flooey

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/6175

monorepo-commit: e08822227848b8c79382a0c4d1e9e98fd704f4ba