Removing a layer of component nesting in ThreadedComments
With a few props/event about to be added to the reply composer of ThreadedComments,
it was a bit annoying having to drill one more level deep for a component that didn't make
a huge difference to be factored out by itself. I am moving ViewerAvatarWithComposer into
the ReplyComponent to remove a level of nesting.

Test Plan:
ThreadedComments still work fine on the testbed - no behavior changes.


Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/6490

monorepo-commit: 8fa21528c8fa8c34de2821bfd6e7fad9ee258456