LiveCursors: move out of experimental and unhide docs
This has been in the works for a long time, let's finally get it out
there. There's a number of things we still want to do, but we believe
the API to be stable now, and it's useful as-is so we can improve in the
future.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/6510

monorepo-commit: c260399c3a6e25790d278f5391ada59348f80d47